feat: Refactor SATD graph visualization and utility functions; add HTML template loading
